IGNITE-28819 Move BinaryContext creation helpers to BinaryUtils
Relocate the machinery for creating a BinaryContext into ignite-binary-api's
BinaryUtils so it can be used by modules that must not depend on ignite-core
(the thin client implementation):

 * TestBinaryContext (+ TestBinaryContextListener) and the useTestBinaryCtx
   test flag moved from IgniteUtils to BinaryUtils.
 * New BinaryUtils.binaryContext(...) overload that takes explicit values
   (ignite instance name, class loader, BinaryConfiguration, affinity fields,
   affinity-field-name provider) instead of an IgniteConfiguration.
   IgniteUtils.binaryContext(IgniteConfiguration) now derives those values and
   delegates to it.
 * affinityFieldName(Class) moved to BinaryUtils; all call sites
   (IgniteUtils, CacheConfiguration, CacheObjectBinaryProcessorImpl, IgniteMock)
   repointed and GridCacheUtils.affinityFieldName removed.
